backend/requirements.txt
# Python依赖包列表
# FastAPI后端所需依赖

fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
groq>=0.4.0
python-dotenv>=1.0.0
websockets>=12.0
websocket-client>=1.6.0
langgraph>=0.0.26
strictjson>=6.1.2
python-multipart>=0.0.6

# LLM相关
litellm>=1.0.0

# Memory持久化（PostgreSQL）
psycopg[binary,pool]>=3.1.0
langgraph-checkpoint-postgres>=1.0.0

# TED相关工具
tavily-python>=0.3.0
git+https://github.com/Xintong120/ted-transcript-extractor.git
